server/requirements.txt
flask
flask-sessions